+++ src/Mesh/MsqIMeshP.hpp
@@ -70,7 +70,7 @@ public:
   /*! Get global ids for given vertices.
    */
   virtual void vertices_get_global_id(const VertexHandle vert_array[],
-				      int gid[],
+				      size_t gid[],
 				      size_t num_vtx,
 				      MsqError& err);
